better touch management if idle screen
I found annoying that each touch event would be passed on the GUI even if my screen was idle. In case of idle screen I think it's better to have one touch to "awake" the gui and a second one to trigger the proper button. That's because on a blank screen you do not know what you are pressing and even which window is getting the touch event. So that's fixed now.